Conform to new nds exports
